requirements.txt
pyaudio
PyYAML>=3.0
requests>=2.7.0
py2exe
ostruct
urllib3
auth
GitPython